Start of network, shortcuts system improvement and corrections
Start of network system (testing), shortcuts system improvement, moving of hud game into correct package, start of chunk improvements ...

+++ sources/game/fr/onsiea/engine/game/world/picking/EnumFacing.java
@@ -3,7 +3,7 @@
  */
 package fr.onsiea.engine.game.world.picking;
 
-import org.joml.Vector3f;
+import org.joml.Vector3i;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -14,14 +14,13 @@ import lombok.Getter;
  */
 public enum EnumFacing
 {
-	NONE(new Vector3f()), RIGHT(new Vector3f(1.0f, 0.0f, 0.0f)), LEFT(new Vector3f(-1.0f, 0.0f, 0.0f)),
-	UP(new Vector3f(0.0f, 1.0f, 0.0f)), DOWN(new Vector3f(0.0f, -1.0f, 0.0f)), FRONT(new Vector3f(0.0f, 0.0f, -1.0f)),
-	BACK(new Vector3f(0.0f, 0.0f, 1.0f));
+	NONE(new Vector3i()), RIGHT(new Vector3i(1, 0, 0)), LEFT(new Vector3i(-1, 0, 0)), UP(new Vector3i(0, 1, 0)),
+	DOWN(new Vector3i(0, -1, 0)), FRONT(new Vector3i(0, 0, -1)), BACK(new Vector3i(0, 0, 1));
 
 	@Getter(AccessLevel.PUBLIC)
-	private Vector3f pointsTo;
+	private Vector3i pointsTo;
 
-	EnumFacing(final Vector3f pointsTo)
+	EnumFacing(final Vector3i pointsTo)
 	{
 		this.pointsTo = pointsTo;
 	}
